backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cloudinary>=1.44.1",
    "fastapi[all]>=0.116.1",
    "google-genai>=1.31.0",
    "python-multipart>=0.0.20",
]
